backend/requirements.txt
asgiref==3.5.2
boto3==1.24.28
botocore==1.27.28
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
future==0.18.2
geocoder==1.38.1
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
ratelim==0.1.6
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.1.2
whitenoise==6.2.0
